robot_software/requirements.txt
action-msgs==1.2.1
actionlib-msgs==4.2.3
# ament-cmake-test==1.3.4
ament-copyright==0.12.6
ament-cppcheck==0.12.6
ament-cpplint==0.12.6
ament-flake8==0.12.6
ament-index-python==1.4.0
ament-lint==0.12.6
ament-lint-cmake==0.12.6
ament-package==0.14.0
ament-pep257==0.12.6
ament-uncrustify==0.12.6
ament-xmllint==0.12.6
builtin-interfaces==1.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
composition-interfaces==1.2.1
contourpy==1.1.0
cv-bridge==3.2.1
cycler==0.11.0
diagnostic-msgs==4.2.3
domain-coordinator==0.10.0
filelock==3.12.2
fonttools==4.40.0
geometry-msgs==4.2.3
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
launch==1.0.4
launch-ros==0.19.4
launch-testing==1.0.4
launch-testing-ros==0.19.4
launch-xml==1.0.4
launch-yaml==1.0.4
lifecycle-msgs==1.2.1
MarkupSafe==2.1.3
matplotlib==3.7.1
message-filters==4.3.3
mpmath==1.3.0
nav-msgs==4.2.3
networkx==3.1
numpy==1.25.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
osrf-pycommon==2.0.2
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
psutil==5.9.5
pyparsing==3.1.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
rcl-interfaces==1.2.1
rclpy==3.3.8
rcutils==5.1.3
requests==2.31.0
rmw-dds-common==1.6.0
ros2action==0.18.6
# ros2bag==0.15.5
ros2cli==0.18.6
ros2component==0.18.6
ros2doctor==0.18.6
ros2interface==0.18.6
ros2launch==0.19.4
ros2lifecycle==0.18.6
ros2multicast==0.18.6
ros2node==0.18.6
ros2param==0.18.6
ros2pkg==0.18.6
ros2run==0.18.6
ros2service==0.18.6
ros2topic==0.18.6
# rosbag2-interfaces==0.15.5
# rosbag2-py==0.15.5
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.4
rosidl-cli==3.1.4
rosidl-cmake==3.1.4
rosidl-generator-c==3.1.4
rosidl-generator-cpp==3.1.4
rosidl-generator-py==0.14.4
rosidl-parser==3.1.4
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.0
rosidl-typesupport-cpp==2.0.0
rosidl-typesupport-fastrtps-c==2.2.0
rosidl-typesupport-fastrtps-cpp==2.2.0
rosidl-typesupport-introspection-c==3.1.4
rosidl-typesupport-introspection-cpp==3.1.4
RPi.GPIO==0.7.1
rpyutils==0.2.1
scipy==1.10.1
seaborn==0.12.2
sensor-msgs==4.2.3
shape-msgs==4.2.3
six==1.16.0
sros2==0.10.4
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.3
stereo-msgs==4.2.3
sympy==1.12
tf2-geometry-msgs==0.25.2
tf2-kdl==0.25.2
tf2-msgs==0.25.2
tf2-py==0.25.2
tf2-ros-py==0.25.2
tf2-tools==0.25.2
theora-image-transport==2.5.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
trajectory-msgs==4.2.3
typing_extensions==4.6.3
tzdata==2023.3
ultralytics==8.0.120
unique-identifier-msgs==2.2.1
urllib3==2.0.3
visualization-msgs==4.2.3
